requirements.txt
grpcio==1.46.0
google-cloud-bigquery==3.2.0
requests
click
